+++ src/lib/components/ui/dropdown-menu/dropdown-menu-checkbox-item.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
 	import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
 	import CheckIcon from "@lucide/svelte/icons/check";
-	import MinusIcon from "@lucide/svelte/icons/minus";
 	import { cn, type WithoutChildrenOrChild } from "$lib/utils.js";
 	import type { Snippet } from "svelte";
 
@@ -30,9 +29,7 @@
 >
 	{#snippet children({ checked: isChecked, indeterminate: isIndeterminate })}
 		<span class="pointer-events-none absolute left-2 flex size-3.5 items-center justify-center">
-			{#if isIndeterminate}
-				<MinusIcon class="size-4" />
-			{:else if isChecked}
+			{#if isChecked || isIndeterminate}
 				<CheckIcon class="size-4" />
 			{/if}
 		</span>